feat: implement casts

+++ Translator/tests/pt/up/fe/specs/clcuda/vectoradd.cl
@@ -1,5 +1,5 @@
-kernel void vectoradd(global int* A, global int* B, global int* C) {
+kernel void vectoradd(global int* A, global float* B, global int* C) {
 	size_t i = get_global_id(0);
 	
-	C[i] = A[i] + B[i];
+	C[i] = A[i] + (int) B[i];
 }
